jdb_mapdb: Address warning by -Wformat-length
With GCC7, -Wformat-length warns like this:

src/jdb/jdb_mapdb.cpp: In static member function ‘static bool Jdb_mapdb::show_tree(Treemap*, Mapping::Pcnt, Mdb_types::Order, unsigned int&, unsigned int)’:
src/jdb/jdb_mapdb.cpp:59:1: warning: ‘%u’ directive output may be truncated writing between 1 and 10 bytes into a region of size 6 [-Wformat-length=]
 Jdb_mapdb::show_tree(Treemap* pages, Mapping::Pcnt offset, Mdb_types::Order base_size,
 ^~~~~~~~~
src/jdb/jdb_mapdb.cpp:59:1: note: directive argument in the range [0ul, 4294967295ul]
src/jdb/jdb_mapdb.cpp:45:68: note: format output between 3 and 12 bytes into a destination of size 6
   snprintf (scratchbuf, 6, "%u%c", unsigned(size), "BKMGTPX"[mult]);
                                                                    ^
Which is indeed the case. However, we're limiting the
range of size to 0 <= size < 1024 before thus this overflow
can actually not happen. To silence the warning I'm casting
to short now instead of unsigned, limiting the data type
appropriately to avoid the warning.

+++ src/jdb/jdb_mapdb.cpp
@@ -41,7 +41,7 @@ size_str (Mword size)
       size >>= 10;
       mult++;
     }
-  snprintf (scratchbuf, 6, "%u%c", unsigned(size), "BKMGTPX"[mult]);
+  snprintf(scratchbuf, 6, "%hu%c", (unsigned short)size, "BKMGTPX"[mult]);
   return scratchbuf;
 }
